Add files via upload
These are the initial files of the project. The ANN, RNN and LSTM cpp files inherent the Block class that would be used by the DynamicNetworkGenerator and the KeyGenerator class is used to generate the ID keys for the neural networks, the AI's and Core's apart from the Main Core of the System.